src/test/java/com/cloudogu/scm/ssh/accesstoken/AccessTokenCommandTest.java
/*
 * Copyright (c) 2020 - present Cloudogu GmbH
 *
 * This program is free software: you can redistribute it and/or modify it under
 * the terms of the GNU Affero General Public License as published by the Free
 * Software Foundation, version 3.
 *
 * This program is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
 * details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program. If not, see https://www.gnu.org/licenses/.
 */

package com.cloudogu.scm.ssh.accesstoken;

import com.cloudogu.scm.ssh.simplecommand.SimpleCommandContext;
import com.fasterxml.jackson.databind.JsonNode;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.apache.shiro.subject.Subject;
import org.apache.shiro.util.ThreadContext;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Answers;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import sonia.scm.config.ScmConfiguration;
import sonia.scm.security.AccessToken;
import sonia.scm.security.AccessTokenBuilder;
import sonia.scm.security.AccessTokenBuilderFactory;

import java.io.ByteArrayOutputStream;
import java.io.IOException;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class AccessTokenCommandTest {

  @Mock
  private Subject subject;
  @Mock
  private ScmConfiguration scmConfiguration;
  @Mock
  private SimpleCommandContext context;
  @Mock
  private AccessTokenBuilderFactory tokenBuilderFactory;
  @Mock(answer = Answers.RETURNS_SELF)
  private AccessTokenBuilder tokenBuilder;
  @Mock
  private AccessToken token;

  @InjectMocks
  private AccessTokenCommand accessTokenCommand;

  @BeforeEach
  void initSubject() {
    ThreadContext.bind(subject);
  }

  @AfterEach
  void tearDownSubject() {
    ThreadContext.unbindSubject();
  }

  @Test
  void shouldReturnValidAccessTokenAndIndex() throws IOException {
    String compactToken = "token123";
    ByteArrayOutputStream baos = new ByteArrayOutputStream();
    when(subject.getPrincipal()).thenReturn("trillian");
    when(scmConfiguration.getBaseUrl()).thenReturn("baseurl/");
    when(context.getOutputStream()).thenReturn(baos);
    when(tokenBuilderFactory.create()).thenReturn(tokenBuilder);
    when(tokenBuilder.build()).thenReturn(token);
    when(token.compact()).thenReturn(compactToken);

    accessTokenCommand.execute(context);

    JsonNode result = new ObjectMapper().readTree(baos.toByteArray());

    assertThat(result.get("accessToken").textValue()).isEqualTo(compactToken);
    assertThat(result.get("_links").get("index").get("href").textValue()).isEqualTo("baseurl/api/v2");
  }
}
